# docs/v1/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "이것은 Project. Stock Exchange API 문서입니다.",
        "title": "Project. Stock Exchange API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://github.com/0ghost0-dev/PJSe/blob/master/license"
        },
        "version": "1.0"
    },
    "host": "localhost:4000",
    "basePath": "/",
    "paths": {
        "/api/v1/admin/activation": {
            "patch": {
                "description": "첫 번째 관리자 계정을 활성화합니다. 이 작업은 primary admin만 수행할 수 있습니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin - Activation"
                ],
                "summary": "첫 번째 관리자 계정 활성화",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Basic {BASE64_ENCODED_CREDENTIALS}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "성공 시 관리자 계정 활성화 메시지 반환",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "접근 거부 또는 권한 없음",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "서버 오류 발생 시 에러 메시지 반환",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}